+++ Runtime.sln
@@ -83,13 +83,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Protobuf", "Specifications\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rudimentary", "Specifications\Rudimentary\Rudimentary.csproj", "{06EF9A15-F2C6-478D-A6F0-630C101658E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Security", "Specifications\Security\Security.csproj", "{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.Clients", "Specifications\Services.Clients\Services.Clients.csproj", "{8AC3B083-931F-469F-BD2C-6F12E98304F1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services", "Specifications\Services\Services.csproj", "{AD4B3DD4-056E-4039-B802-1D696698559C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Versioning", "Specifications\Versioning\Versioning.csproj", "{044DC252-901E-407F-A4F7-F2F968591DE9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Embeddings.Store.Services.Grpc", "Source\Embeddings.Store.Services.Grpc\Embeddings.Store.Services.Grpc.csproj", "{1F21CF7D-0876-4F29-9CC4-BFAE0FD7BA79}"
 EndProject
@@ -138,6 +134,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hosting", "Source\Hosting\Hosting.csproj", "{BCE38FE8-5BAC-47D4-8D65-E705209D2EB1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Source\Domain\Domain.csproj", "{DF39703D-7C80-407F-964E-B685A219058C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Specifications\Domain\Domain.csproj", "{03680B34-81F5-40B7-8D1C-1A5A1188CC64}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -608,18 +606,6 @@ Global
 		{06EF9A15-F2C6-478D-A6F0-630C101658E1}.Release|x64.Build.0 = Release|Any CPU
 		{06EF9A15-F2C6-478D-A6F0-630C101658E1}.Release|x86.ActiveCfg = Release|Any CPU
 		{06EF9A15-F2C6-478D-A6F0-630C101658E1}.Release|x86.Build.0 = Release|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Debug|x64.Build.0 = Debug|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Debug|x86.Build.0 = Debug|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Release|x64.ActiveCfg = Release|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Release|x64.Build.0 = Release|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Release|x86.ActiveCfg = Release|Any CPU
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF}.Release|x86.Build.0 = Release|Any CPU
 		{8AC3B083-931F-469F-BD2C-6F12E98304F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8AC3B083-931F-469F-BD2C-6F12E98304F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8AC3B083-931F-469F-BD2C-6F12E98304F1}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -644,18 +630,6 @@ Global
 		{AD4B3DD4-056E-4039-B802-1D696698559C}.Release|x64.Build.0 = Release|Any CPU
 		{AD4B3DD4-056E-4039-B802-1D696698559C}.Release|x86.ActiveCfg = Release|Any CPU
 		{AD4B3DD4-056E-4039-B802-1D696698559C}.Release|x86.Build.0 = Release|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Debug|x64.Build.0 = Debug|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Debug|x86.Build.0 = Debug|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Release|x64.ActiveCfg = Release|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Release|x64.Build.0 = Release|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Release|x86.ActiveCfg = Release|Any CPU
-		{044DC252-901E-407F-A4F7-F2F968591DE9}.Release|x86.Build.0 = Release|Any CPU
 		{1F21CF7D-0876-4F29-9CC4-BFAE0FD7BA79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1F21CF7D-0876-4F29-9CC4-BFAE0FD7BA79}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1F21CF7D-0876-4F29-9CC4-BFAE0FD7BA79}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -944,6 +918,18 @@ Global
 		{DF39703D-7C80-407F-964E-B685A219058C}.Release|x64.Build.0 = Release|Any CPU
 		{DF39703D-7C80-407F-964E-B685A219058C}.Release|x86.ActiveCfg = Release|Any CPU
 		{DF39703D-7C80-407F-964E-B685A219058C}.Release|x86.Build.0 = Release|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Debug|x64.Build.0 = Debug|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Debug|x86.Build.0 = Debug|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Release|Any CPU.Build.0 = Release|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Release|x64.ActiveCfg = Release|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Release|x64.Build.0 = Release|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Release|x86.ActiveCfg = Release|Any CPU
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{79AB0F5F-4DA5-40C5-9A0B-BE392F16B7C0} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
@@ -984,10 +970,8 @@ Global
 		{2E65C84C-674A-44DA-AD76-190BAD5625D9} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
 		{9B1999BB-0685-45C7-AC3B-2D78D60E99CB} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
 		{06EF9A15-F2C6-478D-A6F0-630C101658E1} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
-		{91D30FCF-91FB-49A3-BC04-F72A6E92D4EF} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
 		{8AC3B083-931F-469F-BD2C-6F12E98304F1} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
 		{AD4B3DD4-056E-4039-B802-1D696698559C} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
-		{044DC252-901E-407F-A4F7-F2F968591DE9} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
 		{1F21CF7D-0876-4F29-9CC4-BFAE0FD7BA79} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 		{A4F26F29-D576-4C8F-8A56-BEDB544F884F} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 		{2AE953FF-3FCD-4D0A-BE5E-F8000FF21FF5} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
@@ -1012,5 +996,6 @@ Global
 		{3AA4AB8B-5820-4D52-A3BD-2958ADD03327} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 		{BCE38FE8-5BAC-47D4-8D65-E705209D2EB1} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
 		{DF39703D-7C80-407F-964E-B685A219058C} = {562DC4C5-91AB-4280-B40C-A2E0C43AA21C}
+		{03680B34-81F5-40B7-8D1C-1A5A1188CC64} = {95D3ED76-BB5F-442F-A99A-F48FF34FEDA6}
 	EndGlobalSection
 EndGlobal
