+++ Specifications/Domain/Domain.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <Import Project="../../specs.props" />
+
+    <PropertyGroup>
+        <AssemblyName>Dolittle.Runtime.Domain.Specs</AssemblyName>
+        <RootNamespace>Dolittle.Runtime.Domain</RootNamespace>
+
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Logging" Version="$(MicrosoftExtensionsVersion)" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="../../Source/Domain/Domain.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Platform" />
+    </ItemGroup>
+
+</Project>
