+++ Specifications/Domain/Platform/for_Version/when_instantiating_without_pre_release_tag.cs
@@ -1,14 +1,13 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using Machine.Specifications;
 
-namespace Dolittle.Runtime.Versioning.for_Version;
+namespace Dolittle.Runtime.Domain.Platform.for_Version;
 
 public class when_instantiating_without_pre_release_tag
 {
-    static Version result;
-    Because of = () => result = new Version(1, 2, 3, 4);
+    static System.Version result;
+    Because of = () => result = new System.Version(1, 2, 3, 4);
     It should_not_be_considered_a_pre_release = () => result.IsPreRelease.ShouldBeFalse();
 }
