+++ Specifications/Execution/for_Claims/when_equating_and_getting_hashcode/two_claims_with_different_claims.cs
@@ -2,14 +2,12 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.Linq;
-using System.Security.Claims;
 using Machine.Specifications;
 
-namespace Dolittle.Runtime.Security.for_Claims.when_equating_and_getting_hashcode;
+namespace Dolittle.Runtime.Execution.for_Claims.when_equating_and_getting_hashcode;
 
 [Subject(typeof(Claims), nameof(Equals))]
-public class two_claims_with_identical_claims_in_different_order
+public class two_claims_with_different_claims
 {
     static Claims first;
     static Claims second;
@@ -21,15 +19,18 @@ public class two_claims_with_identical_claims_in_different_order
 
     Establish context = () =>
     {
-        var list = new List<Claim>
+        var list_one = new List<System.Security.Claims.Claim>
         {
             new("4", "4", "4"),
-            new("1", "1", "1"),
+            new("1", "1", "1")
+        };
+        var list_two = new List<System.Security.Claims.Claim>
+        {
             new("2", "2", "2"),
             new("3", "3", "3")
         };
-        first = new Claims(list.ToArray());
-        second = new Claims(list.OrderBy(_ => _.Name).ToArray());
+        first = new Claims(list_one.ToArray());
+        second = new Claims(list_two.ToArray());
     };
 
     Because of = () =>
@@ -40,8 +41,8 @@ public class two_claims_with_identical_claims_in_different_order
         hash_code_is_equal = first.GetHashCode() == second.GetHashCode();
     };
 
-    It should_be_equal_when_using_the_Equals_method = () => is_equal_by_method.ShouldBeTrue();
-    It should_be_equal_when_using_the_equals_operator = () => is_equal_by_operator.ShouldBeTrue();
-    It should_not_be_not_equal_when_using_the_not_equals_operator = () => is_not_equal_by_operator.ShouldBeFalse();
-    It should_have_the_same_hashcode = () => hash_code_is_equal.ShouldBeTrue();
+    It should_not_be_equal_when_using_the_Equals_method = () => is_equal_by_method.ShouldBeFalse();
+    It should_not_be_equal_when_using_the_equals_operator = () => is_equal_by_operator.ShouldBeFalse();
+    It should_not_be_equal_when_using_the_not_equals_operator = () => is_not_equal_by_operator.ShouldBeTrue();
+    It should_not_have_the_same_hashcode = () => hash_code_is_equal.ShouldBeFalse();
 }
