+++ Specifications/Execution/for_Claims/when_equating_and_getting_hashcode/two_claims_with_identical_claims_in_the_same_order.cs
@@ -2,14 +2,12 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.Linq;
-using System.Security.Claims;
 using Machine.Specifications;
 
-namespace Dolittle.Runtime.Security.for_Claims.when_equating_and_getting_hashcode;
+namespace Dolittle.Runtime.Execution.for_Claims.when_equating_and_getting_hashcode;
 
 [Subject(typeof(Claims), nameof(Equals))]
-public class two_claims_with_different_number_of_claims
+public class two_claims_with_identical_claims_in_the_same_order
 {
     static Claims first;
     static Claims second;
@@ -21,7 +19,7 @@ public class two_claims_with_different_number_of_claims
 
     Establish context = () =>
     {
-        var list = new List<Claim>
+        var list = new List<System.Security.Claims.Claim>
         {
             new("4", "4", "4"),
             new("1", "1", "1"),
@@ -29,7 +27,7 @@ public class two_claims_with_different_number_of_claims
             new("3", "3", "3")
         };
         first = new Claims(list.ToArray());
-        second = new Claims(list.Take(2).ToArray());
+        second = new Claims(list.ToArray());
     };
 
     Because of = () =>
@@ -40,8 +38,8 @@ public class two_claims_with_different_number_of_claims
         hash_code_is_equal = first.GetHashCode() == second.GetHashCode();
     };
 
-    It should_not_be_equal_when_using_the_Equals_method = () => is_equal_by_method.ShouldBeFalse();
-    It should_not_be_equal_when_using_the_equals_operator = () => is_equal_by_operator.ShouldBeFalse();
-    It should_not_be_equal_when_using_the_not_equals_operator = () => is_not_equal_by_operator.ShouldBeTrue();
-    It should_not_have_the_same_hashcode = () => hash_code_is_equal.ShouldBeFalse();
+    It should_be_equal_when_using_the_Equals_method = () => is_equal_by_method.ShouldBeTrue();
+    It should_be_equal_when_using_the_equals_operator = () => is_equal_by_operator.ShouldBeTrue();
+    It should_not_be_not_equal_when_using_the_not_equals_operator = () => is_not_equal_by_operator.ShouldBeFalse();
+    It should_have_the_same_hashcode = () => hash_code_is_equal.ShouldBeTrue();
 }
