--- Specifications/Security/Security.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="../../specs.props" />
-
-  <PropertyGroup>
-    <AssemblyName>Dolittle.Runtime.Security.Specs</AssemblyName>
-  </PropertyGroup>
-
-
-
-</Project>
